move files preparing for math in worker
