+++ src/internal/worker.ts
